# swagger.yml
swagger: "2.0"
info:
  description: "This is the API documentation for the android application WGPlaner."
  version: "0.0.1"
  title: "WGPlaner API Documentation"
  contact:
    name: WGPlaner-Team
    email: "wgplaner@ameyering.de"
  license:
    name: "MIT"
    url: "https://github.com/WGPlaner/wg_planer_server/blob/master/LICENSE"
host: "api.wgplaner.ameyering.de"
basePath: "/"
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  UserIDAuth:
    description: For accessing user related parts of the API a valid userID must be passed in 'Authorization' header.
    type: apiKey
    name: Authorization
    in: header
  FirebaseIDAuth:
    description: For accessing user related parts of the API a valid userID must be passed in 'Authorization' header.
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /groups/{groupId}/create-code:
    get:
      tags:
      - group
      description: Creates and returns a group code. Deletes old codes if they exist.
      operationId: createGroupCode
      security:
        - UserIDAuth: []
      parameters:
        - name: groupId
          in: path
          description: The internal ID of the group
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/GroupCode"
        400:
          description: Invalid groupId
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthroized
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Group not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /groups/join/{groupCode}:
    get:
      tags:
      - group
      description: The authenticated user joins a group through "groupCode"
      operationId: joinGroupHelp
      produces:
        - text/html
      parameters:
        - name: groupCode
          in: path
          description: Group join code
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            type: string
    post:
      tags:
      - group
      description: The authenticated user joins a group through "groupCode"
      operationId: joinGroup
      security:
        - UserIDAuth: []
      parameters:
        - name: groupCode
          in: path
          description: Group join code
          required: true
          type: string
          pattern: '^[A-Z0-9]{12}$'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Group"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /groups/{groupId}:
    get:
      tags:
      - group
      description: Returns the group object
      operationId: getGroup
      security:
        - UserIDAuth: []
      parameters:
        - name: groupId
          in: path
          description: The internal ID of the group
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Group"
        400:
          description: Invalid groupId
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Group not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /groups/leave:
    post:
      tags:
      - group
      description: The authenticated user leaves his group
      operationId: leaveGroup
      security:
        - UserIDAuth: []
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /groups:
    post:
      tags:
      - group
      description: Creates a new group
      operationId: createGroup
      security:
        - UserIDAuth: []
      parameters:
        - in: body
          name: body
          description: The group to create (contains displayName, etc)
          required: true
          schema:
            $ref: "#/definitions/Group"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Group"
        400:
          description: Invalid group data
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized User
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /groups/{groupId}/image:
    get:
      tags:
      - group
      description: Get the group's profile image
      operationId: getGroupImage
      security:
        - UserIDAuth: []
      parameters:
        - name: groupId
          in: path
          description: The internal ID of the group
          required: true
          type: string
          format: uuid
      produces:
        - application/octet-stream
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: string
            format: binary
            readOnly: true
        400:
          description: Invalid groupId
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
      - group
      description: Update the group's profile image
      operationId: updateGroupImage
      security:
        - UserIDAuth: []
      parameters:
        - name: groupId
          in: path
          description: The internal ID of the group
          required: true
          type: string
          format: uuid
        - name: profileImage
          in: formData
          description: Success
          required: true
          type: file
      consumes:
        - multipart/form-data
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        400:
          description: Invalid groupId
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Group:
    required:
      - displayName
    properties:
      uid:
        type: string
        format: uuid
        readOnly: true
      displayName:
        type: string
      currency:
        type: string
        maxLength: 4
      photoUrl:
        type: string
        format: uri
      admins:
        type: array
        items:
          type: string
        readOnly: true
      members:
        type: array
        items:
          type: string
        readOnly: true
      createdAt:
        type: string
        format: date-time
        readOnly: true
      updatedAt:
        type: string
        format: date-time
        readOnly: true
  GroupCode:
    required:
      - groupUid
      - code
      - validUntil
    properties:
      groupUid:
        type: string
        format: uuid
      code:
        type: string
        pattern: '^[A-Z0-9]{12}$'
      validUntil:
        type: string
        format: date-time
        readOnly: true
...
  ErrorResponse:
    required:
      - status
      - message
    properties:
      status:
        type: integer
      message:
        type: string
  SuccessResponse:
    required:
      - status
      - message
    properties:
      status:
        type: integer
      message:
        type: string